+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Changelog
 
+## 1.0.0 (03/05/2020)
+
+### Breaking changes
+ * Migration to `@animoca/ethereum-contracts-core_library:1.0.0` with `solc:0.6.x` and `@openzeppelin/contracts:3.x`.
+
+### New features
+ * Added `NftStakingMock.sol` and `NftStakingTestableMock.sol`.
+
+### Improvements
+ * Better abstraction of core staking features.
+
 ## 0.0.1 (15/04/2020)
 * Initial commit.